Move all the popup stuff into the popups directory
