+++ src/components/infopopups/infopopup.css
